Feat: Implement wishlist functionality with EmptyWishList and ItemWishlist components, and update related models and navigation
