--- app/src/main/java/com/datn/viettech_md_12/screen/WishlistScreen.kt
@@ -1,366 +1,0 @@
-package com.datn.viettech_md_12.screen
-
-import FavoriteItem
-import MyButton
-import android.content.Context
-import android.util.Log
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.DismissDirection
-import androidx.compose.material.DismissValue
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Icon
-import androidx.compose.material.SnackbarDuration
-import androidx.compose.material.SnackbarHost
-import androidx.compose.material.SnackbarHostState
-import androidx.compose.material.SnackbarResult
-import androidx.compose.material.SwipeToDismiss
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Remove
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material.Text
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.ArrowBackIosNew
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.outlined.Delete
-import androidx.compose.material.rememberDismissState
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SmallTopAppBar
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarColors
-import androidx.compose.material3.TopAppBarDefaults
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.tooling.preview.Devices
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
-import coil.compose.AsyncImage
-import com.datn.viettech_md_12.R
-import com.datn.viettech_md_12.component.cart_component.CartNotLogin
-import com.datn.viettech_md_12.data.model.ProductModel
-import com.datn.viettech_md_12.viewmodel.ProductViewModel
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import java.text.NumberFormat
-import java.util.Locale
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun WishlistScreen(viewModel: ProductViewModel, navController: NavController) {
-    val snackbarHostState = remember { SnackbarHostState() }
-    val favoriteProducts by viewModel.favoriteProducts.collectAsState(initial = emptyList())
-    val context = LocalContext.current
-    val coroutineScope = rememberCoroutineScope()
-
-    var deletedItem by remember { mutableStateOf<FavoriteItem?>(null) }
-    val sharedPreferences = context.getSharedPreferences("MyPrefs", Context.MODE_PRIVATE)
-    val accessToken = sharedPreferences.getString("accessToken", null)
-    LaunchedEffect(Unit) {
-        viewModel.getFavoriteProducts(context)
-    }
-
-    Scaffold(
-        modifier = Modifier
-            .fillMaxSize()
-            .systemBarsPadding(),
-        topBar = {
-            SmallTopAppBar(
-                title = {
-                   Text(text = "Yêu thích", fontWeight = FontWeight.Bold, fontSize = 20.sp, color = Color.Black)
-                },
-                colors = TopAppBarColors(
-                    containerColor = Color.White,
-                    scrolledContainerColor = Color.Transparent,
-                    navigationIconContentColor = Color.Black,
-                    titleContentColor = Color.Black,
-                    actionIconContentColor = Color.Transparent
-                ),
-                navigationIcon = {
-                    IconButton(onClick = { navController.popBackStack() }) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
-                    }
-                },
-                modifier = Modifier.shadow(elevation = 2.dp),
-            )
-        },
-        snackbarHost = {
-            SnackbarHost(snackbarHostState)
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.White)
-                .padding(innerPadding)
-        ) {
-            LazyColumn(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-                if (favoriteProducts.isNotEmpty()) {
-                    items(favoriteProducts, key = { it.product.id }) { product ->
-                        ItemFavorite(
-                            favoriteItem = product,
-                            onItemDismissed = { item ->
-                                deletedItem = item
-                                viewModel.removeFromFavorites(item.product.id, context)
-
-//                                coroutineScope.launch {
-//                                    val result = snackbarHostState.showSnackbar(
-//                                        message = "Đã xóa khỏi yêu thích",
-//                                        actionLabel = "Khôi phục",
-//                                        duration = SnackbarDuration.Long
-//                                    )
-//                                    if (result == SnackbarResult.ActionPerformed) {
-//                                        // Gọi addToFavorites trước
-//                                        viewModel.addToFavorites(item.product.id, context)
-//
-//                                        // Sau đó cập nhật lại danh sách — dùng delay nhẹ để đợi API hoàn tất
-//                                        delay(500) // Có thể thay đổi tùy tốc độ server
-//                                        viewModel.getFavoriteProducts(context)
-//                                    }
-//                                }
-                            },
-
-                        )
-                    }
-
-                } else if (accessToken.isNullOrEmpty()) {
-                    item { CartNotLogin(navController) }
-                } else{
-                    item { EmptyWishList(navController) }
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@Composable
-fun ItemFavorite(
-    favoriteItem: FavoriteItem,
-    onItemDismissed: (FavoriteItem) -> Unit
-) {
-    val product = favoriteItem.product
-    val BASE_URL = "http://103.166.184.249:3056/"
-    val itemPriceFormatted = NumberFormat.getCurrencyInstance(Locale("vi", "VN")).format(product.product_price)
-
-    var showDeleteConfirm by remember { mutableStateOf(false) } // xac nhan xoa
-    val coroutineScope = rememberCoroutineScope()
-    val dismissState = rememberDismissState(
-        confirmStateChange = {
-            true // Cho phép dismiss hoàn tất
-        }
-    )
-
-   //lang nghe trang thai khi vuot
-    LaunchedEffect(dismissState.currentValue) {
-        if (dismissState.currentValue == DismissValue.DismissedToStart) {
-            showDeleteConfirm = true
-        } // xac nhan xoa
-    }
-
-    SwipeToDismiss(
-        state = dismissState,
-        background = {
-            if (showDeleteConfirm) {
-                // Giao diện xác nhận xoá
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(16.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(Color(0xFFEF4139)),
-                    contentAlignment = Alignment.CenterEnd
-                ) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Text(
-                            text = "Xác nhận xoá?",
-                            color = Color.White,
-                            modifier = Modifier.padding(end = 8.dp),
-                            fontSize = 14.sp
-                        )
-                        TextButton(
-                            onClick = {
-                                onItemDismissed(favoriteItem)
-                                showDeleteConfirm = false
-                                coroutineScope.launch {
-                                    dismissState.reset()
-                                }
-                            }
-                        ) {
-                            Text("Xoá", fontWeight = FontWeight.Bold, color = Color.White)
-                        }
-                    }
-                }
-            } else {
-                // Nền mặc định khi chưa vuốt
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(16.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(Color(0xFFEF4139)),
-                    contentAlignment = Alignment.CenterEnd
-                ){
-                    Icon(
-                        imageVector = Icons.Default.Delete,
-                        contentDescription = "Xóa",
-                        tint = Color.White,
-                        modifier = Modifier.size(28.dp)
-                    )
-                }
-            }
-        },
-        directions = setOf(DismissDirection.EndToStart)
-    ) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(12.dp)),
-            colors = CardDefaults.cardColors(containerColor = Color.White)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(12.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                AsyncImage(
-                    model = "$BASE_URL${product.product_thumbnail}",
-                    contentDescription = null,
-                    modifier = Modifier
-                        .size(70.dp)
-                        .clip(RoundedCornerShape(10.dp))
-                        .background(Color(0xFFF4F4F4))
-                )
-
-                Spacer(modifier = Modifier.width(12.dp))
-
-                Column(modifier = Modifier.weight(1f)) {
-                    Text(
-                        text = product.product_name,
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 16.sp,
-                        color = Color.Black
-                    )
-                    Spacer(modifier = Modifier.height(4.dp))
-                    Text(
-                        text = itemPriceFormatted,
-                        color = Color(0xFFF11F10),
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 15.sp
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun EmptyWishList(
-    navController: NavController,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.White)
-            .padding(horizontal = 20.dp)
-            .systemBarsPadding(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        //ảnh giỏ hàng
-        Spacer(Modifier.height(10.dp))
-        Box(
-            modifier = Modifier
-                .size(340.dp)
-                .clip(shape = RoundedCornerShape(16.dp))
-                .background(color = Color(0xFFF4FDFA)),
-            contentAlignment = Alignment.Center,
-        ) {
-            Image(
-                modifier = Modifier
-                    .size(200.dp),
-                painter = painterResource(R.drawable.img_empty_wishlist),
-                contentDescription = "empty cart image",
-                contentScale = ContentScale.Crop
-            )
-        }
-        Spacer(Modifier.height(30.dp))
-        //Text
-        Text(
-            "Danh sách yêu thích của bạn đang trống",
-            color = Color.Black,
-            fontSize = 18.sp,
-            fontWeight = FontWeight.Bold,
-            textAlign = TextAlign.Center
-        )
-        Spacer(Modifier.height(10.dp))
-        Text(
-            "Nhấn vào nút trái tim để bắt đầu lưu các mục yêu thích của bạn.",
-            color = Color.Gray,
-            fontSize = 14.sp,
-            textAlign = TextAlign.Center,
-        )
-        Spacer(Modifier.height(30.dp))
-        //Button
-        MyButton(
-            text = "Khám phá",
-            onClick = {navController.navigate("home") },
-            modifier = Modifier,
-            backgroundColor = Color(0xFF21D4B4),
-            textColor = Color.White,
-        )
-    }
-}
-
-@Preview(showBackground = true, showSystemUi = true, device = Devices.PIXEL_7)
-@Composable
-fun PreviewWishList() {
-//    EmptyWishList()
-}
