+++ app/src/main/java/com/datn/viettech_md_12/screen/wishlist/WishlistScreen.kt
@@ -1,0 +1,117 @@
+package com.datn.viettech_md_12.screen.wishlist
+
+import WishlistItem
+import android.content.Context
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarColors
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
+import com.datn.viettech_md_12.component.cart_component.CartNotLogin
+import com.datn.viettech_md_12.viewmodel.ProductViewModel
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun WishlistScreen(viewModel: ProductViewModel, navController: NavController) {
+    val snackbarHostState = remember { SnackbarHostState() }
+    val favoriteProducts by viewModel.favoriteProducts.collectAsState(initial = emptyList())
+    val context = LocalContext.current
+
+    var deletedItem by remember { mutableStateOf<WishlistItem?>(null) }
+    val sharedPreferences = context.getSharedPreferences("MyPrefs", Context.MODE_PRIVATE)
+    val accessToken = sharedPreferences.getString("accessToken", null)
+
+    LaunchedEffect(Unit) {
+        viewModel.getFavoriteProducts(context)
+    }
+
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize()
+            .systemBarsPadding(),
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(
+                        text = "Yêu thích",
+                        fontWeight = FontWeight.Bold,
+                        fontSize = 20.sp,
+                        color = Color.Black
+                    )
+                },
+                modifier = Modifier.shadow(elevation = 2.dp),
+                navigationIcon = {
+                    IconButton(onClick = { navController.popBackStack() }) {
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                    }
+                },
+                colors = TopAppBarColors(
+                    containerColor = Color.White,
+                    scrolledContainerColor = Color.Transparent,
+                    navigationIconContentColor = Color.Black,
+                    titleContentColor = Color.Black,
+                    actionIconContentColor = Color.Transparent
+                )
+            )
+        },
+        snackbarHost = {
+            SnackbarHost(snackbarHostState)
+        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(Color.White)
+                .padding(innerPadding)
+        ) {
+            LazyColumn(verticalArrangement = Arrangement.spacedBy(12.dp)) {
+                if (favoriteProducts.isNotEmpty()) {
+                    items(favoriteProducts, key = { it.product.id }) { product ->
+                        ItemWishlist(
+                            wishlistItem = product,
+                            onItemDismissed = { item ->
+                                deletedItem = item
+                                viewModel.removeFromFavorites(item.product.id, context)
+                            },
+                        )
+                    }
+
+                } else if (accessToken.isNullOrEmpty()) {
+                    item { CartNotLogin(navController) }
+                } else {
+                    item { EmptyWishList(navController) }
+                }
+            }
+        }
+    }
+}
+
